// src/HOPPER.Frontend/src/app/api/model/modrinthSearchHitDto.ts
/**
 * HOPPER.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ManifestModDtoSize } from './manifestModDtoSize';


export interface ModrinthSearchHitDto { 
    projectId: string;
    slug?: string | null;
    title: string;
    description?: string | null;
    author?: string | null;
    iconUrl?: string | null;
    downloads: ManifestModDtoSize;
    follows: ManifestModDtoSize;
    categories: Array<string>;
    dateModified?: string | null;
    installed: boolean;
}